make tests unit tests
