+++ dockstore-webservice/src/test/java/io/dockstore/webservice/helpers/SourceCodeRepoFactoryTest.java
@@ -18,8 +18,13 @@ package io.dockstore.webservice.helpers;
 
 import java.util.Map;
 import java.util.Optional;
+
+import io.dockstore.webservice.core.BioWorkflow;
+import io.dockstore.webservice.core.Tool;
 import org.junit.Assert;
 import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
 
 /**
  * @author dyuen
@@ -79,4 +84,151 @@ public class SourceCodeRepoFactoryTest {
 
     }
 
+    /**
+     * Tests that parsing a GitLab URL works
+     */
+    @Test
+    public void testGitLabUrlRegexParsing() {
+        GitLabSourceCodeRepo repo = new GitLabSourceCodeRepo("fakeUser", "fakeToken");
+        final BioWorkflow entry = new BioWorkflow();
+
+        /* Test good URLs */
+        entry.setGitUrl("git@gitlab.com:dockstore/dockstore-ui.git");
+        String gitlabId = repo.getRepositoryId(entry);
+        assertEquals("gitlab ID parse check", "dockstore/dockstore-ui", gitlabId);
+
+        entry.setGitUrl("git@gitlab.com:dockstore-cow/goat.git");
+        gitlabId = repo.getRepositoryId(entry);
+        assertEquals("gitlab ID parse check", "dockstore-cow/goat", gitlabId);
+
+        entry.setGitUrl("git@gitlab.com:dockstore.dot/goat.bat.git");
+        gitlabId = repo.getRepositoryId(entry);
+        assertEquals("gitlab ID parse check", "dockstore.dot/goat.bat", gitlabId);
+
+        entry.setGitUrl("git@gitlab.com:dockstore.dot/goat.git");
+        gitlabId = repo.getRepositoryId(entry);
+        assertEquals("gitlab ID parse check", "dockstore.dot/goat", gitlabId);
+
+        /* Test bad URLs */
+        entry.setGitUrl("git@gitlab.com/dockstore/dockstore-ui.git");
+        gitlabId = repo.getRepositoryId(entry);
+        assertEquals("gitlab ID parse check", null, gitlabId);
+
+        entry.setGitUrl("git@gitlab.com:dockstore:dockstore-ui.git");
+        gitlabId = repo.getRepositoryId(entry);
+        assertEquals("gitlab ID parse check", null, gitlabId);
+
+        entry.setGitUrl("git@gitlab.com:/dockstore-ui.git");
+        gitlabId = repo.getRepositoryId(entry);
+        assertEquals("gitlab ID parse check", null, gitlabId);
+
+        entry.setGitUrl("git@gitlab.com:dockstore");
+        gitlabId = repo.getRepositoryId(entry);
+        assertEquals("gitlab ID parse check", null, gitlabId);
+
+        entry.setGitUrl("git@gitlab.com:dockstore/dockstore-ui");
+        gitlabId = repo.getRepositoryId(entry);
+        assertEquals("gitlab ID parse check", null, gitlabId);
+
+    }
+
+    /**
+     * Tests that parsing a GitHub URL works
+     */
+    @Test
+    public void testGitHubUrlRegexParsing() {
+        GitHubSourceCodeRepo repo = new GitHubSourceCodeRepo("fakeUser", "fakeToken");
+        final Tool entry = new Tool();
+
+        /* Test good URLs */
+        entry.setGitUrl("git@github.com:dockstore/dockstore-ui.git");
+        String githubId = repo.getRepositoryId(entry);
+        assertEquals("GitHub ID parse check", "dockstore/dockstore-ui", githubId);
+
+        entry.setGitUrl("git@github.com:dockstore-cow/goat.git");
+        githubId = repo.getRepositoryId(entry);
+        assertEquals("GitHub ID parse check", "dockstore-cow/goat", githubId);
+
+        entry.setGitUrl("git@github.com:dockstore.dot/goat.bat.git");
+        githubId = repo.getRepositoryId(entry);
+        assertEquals("GitHub ID parse check", "dockstore.dot/goat.bat", githubId);
+
+        entry.setGitUrl("git@github.com:dockstore.dot/goat.git");
+        githubId = repo.getRepositoryId(entry);
+        assertEquals("GitHub ID parse check", "dockstore.dot/goat", githubId);
+
+        /* Test bad URLs */
+        entry.setGitUrl("git@github.com/dockstore/dockstore-ui.git");
+        githubId = repo.getRepositoryId(entry);
+        assertEquals("GitHub ID parse check", null, githubId);
+
+        entry.setGitUrl("git@github.com:dockstore:dockstore-ui.git");
+        githubId = repo.getRepositoryId(entry);
+        assertEquals("GitHub ID parse check", null, githubId);
+
+        entry.setGitUrl("git@github.com:/dockstore-ui.git");
+        githubId = repo.getRepositoryId(entry);
+        assertEquals("GitHub ID parse check", null, githubId);
+
+        entry.setGitUrl("git@github.com:dockstore");
+        githubId = repo.getRepositoryId(entry);
+        assertEquals("GitHub ID parse check", null, githubId);
+
+        entry.setGitUrl("git@github.com:dockstore/dockstore-ui");
+        githubId = repo.getRepositoryId(entry);
+        assertEquals("GitHub ID parse check", null, githubId);
+
+    }
+
+    /**
+     * Tests that parsing a BitBucket URL works
+     */
+    @Test
+    public void testBitBucketUrlParsing() {
+        BitBucketSourceCodeRepo repo = new BitBucketSourceCodeRepo("fakeUser", "fakeToken");
+        final BioWorkflow entry = new BioWorkflow();
+
+        /* Test good URLs */
+        entry.setGitUrl("git@bitbucket.org:dockstore/dockstore-ui.git");
+        String bitBucketId = repo.getRepositoryId(entry);
+        assertEquals("Bitbucket ID parse check", "dockstore/dockstore-ui", bitBucketId);
+
+        entry.setGitUrl("git@bitbucket.org:dockstore/dockstore-ui.git");
+        bitBucketId = repo.getRepositoryId(entry);
+        assertEquals("Bitbucket ID parse check", "dockstore/dockstore-ui", bitBucketId);
+
+        entry.setGitUrl("git@bitbucket.org:dockstore-cow/goat.git");
+        bitBucketId = repo.getRepositoryId(entry);
+        assertEquals("Bitbucket ID parse check", "dockstore-cow/goat", bitBucketId);
+
+        entry.setGitUrl("git@bitbucket.org:dockstore.dot/goat.bat.git");
+        bitBucketId = repo.getRepositoryId(entry);
+        assertEquals("Bitbucket ID parse check", "dockstore.dot/goat.bat", bitBucketId);
+
+        entry.setGitUrl("git@bitbucket.org:dockstore.dot/goat.git");
+        bitBucketId = repo.getRepositoryId(entry);
+        assertEquals("Bitbucket ID parse check", "dockstore.dot/goat", bitBucketId);
+
+        /* Test bad URLs */
+        entry.setGitUrl("git@bitbucket.org/dockstore/dockstore-ui.git");
+        bitBucketId = repo.getRepositoryId(entry);
+        assertEquals("Bitbucket ID parse check", null, bitBucketId);
+
+        entry.setGitUrl("git@bitbucket.org:dockstore:dockstore-ui.git");
+        bitBucketId = repo.getRepositoryId(entry);
+        assertEquals("Bitbucket ID parse check", null, bitBucketId);
+
+        entry.setGitUrl("git@bitbucket.org:/dockstore-ui.git");
+        bitBucketId = repo.getRepositoryId(entry);
+        assertEquals("Bitbucket ID parse check", null, bitBucketId);
+
+        entry.setGitUrl("git@bitbucket.org:dockstore");
+        bitBucketId = repo.getRepositoryId(entry);
+        assertEquals("Bitbucket ID parse check", null, bitBucketId);
+
+        entry.setGitUrl("git@bitbucket.org:dockstore/dockstore-ui");
+        bitBucketId = repo.getRepositoryId(entry);
+        assertEquals("Bitbucket ID parse check", null, bitBucketId);
+
+    }
 }
